refactor!: tactical nuke INCOMING
