+++ lux.toml
@@ -1,6 +1,6 @@
-package = "rocks.nvim"
-version = "0.1.0"
-lua = "=5.1"
+package = "lux.nvim"
+version = "3.0.0"
+lua = "5.1"
 
 [description]
 summary = "🌒 Neovim plugin management inspired by Cargo, powered by luarocks"
@@ -8,11 +8,11 @@ labels = [ "lua", "luarocks", "neovim", "neovim-plugin", "nvim", "package-manage
 license = "GPL-3.0"
 
 [dependencies]
-toml-edit = "0.6.1-1"
 "fidget.nvim" = "1.6.0-1"
 fzy = "1.0.3-1"
 nvim-nio = "1.10.1-1"
 "rtp.nvim" = "1.2.0-1"
+"pathlib.nvim" = "2.2.3-1"
 
 [test_dependencies]
 nlua = "0.3.0-1"
