+++ plugin/lux-nvim.lua
@@ -1,0 +1,5 @@
+if vim.g.loaded_lux_nvim then
+    return
+end
+
+vim.g.loaded_lux_nvim = true
